HTTP Checkup | Web Response Analyzer
An automated Python script that sends HTTP and HTTPS requests to a list of domains. By examining response codes and sizes for both protocols, this tool gives information about the condition and functionality of web services. The gathered information is saved in a CSV file, making it simple to track and evaluate how responsive a website is. With WebResponseAnalyzer, you can automate your online testing and monitoring chores.

You are allowed to change the description as necessary to appropriately reflect the goal and functionality of your project.